% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2text.R
\name{dje2text}
\alias{dje2text}
\title{Converts a list of dje pdf files to text}
\usage{
dje2text(files = NULL, path = NULL, ex_dir = "dje_txt/", ...)
}
\arguments{
\item{files}{list of files to convert. defaults to NULL, when path should be used.}

\item{path}{directory containing pdf files (and only them) to convert. works with folders.}

\item{ex_dir}{folder to save the files}

\item{...}{other param passed to pdf2text}
}
\description{
Converts a list of dje pdf files to text
}
