% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{download_skeleton_dje}
\alias{download_skeleton_dje}
\title{Downloads DJE structure}
\usage{
download_skeleton_dje(dt_diario)
}
\arguments{
\item{dt_diario}{vector of dates in YYYY-MM-DD format.}
}
\value{
tibble containing all section names and pages. Only for TJSP.
}
\description{
Downloads DJE structure from TJSP site
}
