% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dje.R
\name{parse_dje_tjsp}
\alias{parse_dje_tjsp}
\title{Parse DJE file}
\usage{
parse_dje_tjsp(text_file)
}
\arguments{
\item{text_file}{path of the file to parse.}
}
\value{
tibble containing lawsuit and court information
}
\description{
parse TJSP file obtained from \code{\link[=dje_to_text]{dje_to_text()}}.
}
