% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2text.R
\name{pdf2text}
\alias{pdf2text}
\title{Converts a pdf file to text}
\usage{
pdf2text(a, first_pg = NA, last_pg = NA, r = F, keep_file = T,
  return_string = F, new_file = "repo.txt")
}
\arguments{
\item{a}{pdf file to convert}

\item{first_pg}{page where the convertion should start. defaults to NA, equivalent to 1.}

\item{last_pg}{page where the convertion should end. defaults to NA, equivalent to the last page.}

\item{r}{should the convertiong uses poppler's -raw?}

\item{keep_file}{should we keep the file after convertion?}

\item{return_string}{should this functions return a string?}

\item{new_file}{name of the output file}
}
\description{
Wraps a call to poppler's pdftotext function.
}
