% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2text.R
\name{pdf_to_text}
\alias{pdf_to_text}
\title{Converts a PDF file to text (wraps a call to Poppler's \code{pdftotext} function)}
\usage{
pdf_to_text(file, new_file = NULL, start_pg = NULL, end_pg = NULL,
  raw = FALSE, return_text = FALSE)
}
\arguments{
\item{file}{Path to file}

\item{new_file}{Path to file where to save converted text (not used if
\code{return_text = TRUE})}

\item{start_pg}{Page where conversion should start (defaults to \code{NULL},
equivalent to first page)}

\item{end_pg}{Page where conversion should end (defaults to \code{NULL},
equivalent to last page)}

\item{raw}{Whether conversion should use Poppler's \code{-raw}}

\item{return_text}{If \code{TRUE} this function will return the converted
text as a character vector, and if \code{FALSE} (the default) it will
return the path to a \code{.txt} file containing the converted text}
}
\value{
A character vector either with the converted text or with
the path to the new file
}
\description{
Converts a PDF file to text (wraps a call to Poppler's \code{pdftotext} function)
}
