% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_text.R
\name{dje_to_text}
\alias{dje_to_text}
\title{Converts a directory of DJE PDFs to text (using \code{\link[=pdf_to_text]{pdf_to_text()}})}
\usage{
dje_to_text(path = ".", new_path = NULL, ...)
}
\arguments{
\item{path}{Path to directory containing files to be converted}

\item{new_path}{Path to directory where to save converted files (defaults
to \code{NULL}, simply appending "_txt" to \code{path})}

\item{...}{Parameters passed on to \code{\link[=pdf_to_text]{pdf_to_text()}}}
}
\value{
A character vector with the paths the new files
}
\description{
Converts a directory of DJE PDFs to text (using \code{\link[=pdf_to_text]{pdf_to_text()}})
}
\seealso{
\code{\link[=pdf_to_text]{pdf_to_text()}}
}
